+++ src/views/app/detail-panel/edit/editor/parts/edit-view-send-buttons.tsx
@@ -8,23 +8,28 @@ import React, { FC, useCallback, useMemo } from 'react';
 import { Button, MultiButton, Tooltip, useModal } from '@zextras/carbonio-design-system';
 import { t, useUserSettings } from '@zextras/carbonio-shell-ui';
 
+import { useEditorSend } from 'store/editor';
 import { SendLaterModal } from 'views/app/detail-panel/edit/editor/parts/send-later-modal';
 
 export type EditViewSendButtonsProps = {
 	onSendLater: (autoSendTime: number) => void;
 	onSendNow: () => void;
-	disabled: boolean;
-	tooltip: string;
+	editorId: string;
 };
 
 export const EditViewSendButtons: FC<EditViewSendButtonsProps> = ({
 	onSendLater,
 	onSendNow,
-	disabled,
-	tooltip
+	editorId
 }) => {
 	const { attrs } = useUserSettings();
 	const { createModal, closeModal } = useModal();
+
+	const { status: sendAllowedStatus } = useEditorSend(editorId);
+
+	const disabled = !sendAllowedStatus?.allowed;
+
+	const tooltip = sendAllowedStatus?.reason;
 
 	const onSendLaterClick = useCallback(() => {
 		const modalId = Date.now().toString();
@@ -80,9 +85,6 @@ export const EditViewSendButtons: FC<EditViewSendButtonsProps> = ({
 						onClick={onSendNow}
 						disabledPrimary={disabled}
 						disabledSecondary={disabled}
-						// TOFIX: remove this ts-ignore once SHELL 5.3.0 is released
-						// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-						// @ts-ignore
 						icon={'ChevronDownOutline'}
 						items={multiBtnActions}
 					/>